Only generate positive numerical literals
The parser currently parses `-123` or `-123.456` as unary negation
followed by a literal, but `genExpr` happily generated negative
literals, which never parse back to their original representation.

It might be better to parse negation as part of literals though. In
fact, I suspect the current behavior might be causing the shenanigans
with parsing `-` and `+`!